Added charts & models

+++ database/migrations/2024_05_12_101424_create_transactions_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('type');
             $table->integer('quantity');
 
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')
                 ->references('id')
                 ->on('products')
